+++ admin/manage-users.php
@@ -93,36 +93,42 @@
         </aside>
         <main>
             <h2>Manage Users</h2>
-            <table>
-                <thead>
-                    <tr>
-                        <th>Name</th>
-                        <th>Username</th>
-                        <th>Edit</th>
-                        <th>Delete</th>
-                        <th>Privilege</th>
-                    </tr>
-                </thead>
-                <tbody>
-                    
-                    <?php while($user = mysqli_fetch_assoc($users)) : ?>
-                        <tr>
-                            <td><?= "{$user['firstname']} {$user['lastname']}" ?></td>
-                            <td><?= "{$user['username']}"?></td>
-                            <td><a href="<?= ROOT_URL ?>admin/edit-user.php?id=<?=$user['id']?>" class="btn sm">Edit</a></td>
-                            <td><a href="<?= ROOT_URL ?>admin/delete-user.php?id=<?=$user['id']?>" class="btn sm danger">Delete</a></td>
-                            <td>
-                                <?php if($user['is_admin']== 1):?>
-                                    Admin
-                                <?php elseif($user['is_admin']== 0):?>
-                                    Author
-                                <?php endif?>
-                            </td>
-                        </tr>
-                    <?php endwhile?>
 
-                </tbody>
-            </table>
+            <?php if(mysqli_num_rows($users) > 0) : ?>
+                
+                <table>
+                    <thead>
+                        <tr>
+                            <th>Name</th>
+                            <th>Username</th>
+                            <th>Edit</th>
+                            <th>Delete</th>
+                            <th>Privilege</th>
+                        </tr>
+                    </thead>
+                    <tbody>
+
+                        <?php while($user = mysqli_fetch_assoc($users)) : ?>
+                            <tr>
+                                <td><?= "{$user['firstname']} {$user['lastname']}" ?></td>
+                                <td><?= "{$user['username']}"?></td>
+                                <td><a href="<?= ROOT_URL ?>admin/edit-user.php?id=<?=$user['id']?>" class="btn sm">Edit</a></td>
+                                <td><a href="<?= ROOT_URL ?>admin/delete-user.php?id=<?=$user['id']?>" class="btn sm danger">Delete</a></td>
+                                <td>
+                                    <?php if($user['is_admin']== 1):?>
+                                        Admin
+                                    <?php elseif($user['is_admin']== 0):?>
+                                        Author
+                                    <?php endif?>
+                                </td>
+                            </tr>
+                        <?php endwhile?>
+
+                    </tbody>
+                </table>
+            <?php else : ?>
+                <div class="alert_message error"> No Users found<div>
+            <?php endif ?>
         </main>
     </div>
 </section>
